user-creation-consumer/requirements.txt
pika==1.3.2
python-dotenv==1.0.0
mysql-connector-python==8.0.33